+++ src/components/navigation/navbar/desktop/student/courses/CourseMenuBody.vue
@@ -1,7 +1,7 @@
 <template>
     <section class="flex flex-col lg:flex-row px-8 py-6 border-b -mx-4">
-        <ul class="w-full lg:w-1/2 px-2">
-            <li class="mb-4">
+        <ul class="w-full px-2">
+            <li class="">
                 <menu-item
                     id="nav_desktop_student_menu_courses_all"
                     title="All Courses"
@@ -9,43 +9,6 @@
                     icon-class="fa-globe-americas"
                     target-route-name="student.courses"
                 />
-            </li>
-            <li class="mb-4">
-                <menu-item
-                    title="Favorite Courses"
-                    description="All courses you marked as favorite"
-                    icon-class="fa-star"
-                    target-route-name="home"
-                />
-            </li>
-            <li class="mb-4">
-                <menu-item
-                    title="Active Courses"
-                    description="Courses you are currently enrolled in"
-                    icon-class="fa-users"
-                    target-route-name="home"
-                />
-            </li>
-        </ul>
-        <ul class="w-full lg:w-1/2 px-4">
-            <li class="mb-4">
-                <menu-item
-                    title="My Courses"
-                    description="All courses you ever participated in"
-                    icon-class="fa-meteor"
-                    target-route-name="home"
-                />
-            </li>
-            <li class="mb-4">
-                <menu-item
-                    title="Organisational Stuff"
-                    description="Probably meant to click this one"
-                    icon-class="fa-sitemap"
-                    target-route-name="home"
-                />
-            </li>
-            <li class="mb-4">
-                <menu-item title="Play Chess" description="Check mate" icon-class="fa-chess-knight" target-route-name="home" />
             </li>
         </ul>
     </section>